requirements.txt
google-adk
dnspython
requests